Initial Commit -- My Version of Birthday GIFt site
- Background image
-background size
- I actually had a hard time downloading the gifs 
- flex and flex-direction
- gradients
- hover effects

WORK NEEDED:
I still need to fix a width and height problem on the hover effects of the gifs. They are not containing the whole div. I tried several other proprieties like "cover" or max and min width and height. None of it works and all the images sizes keep changing regardless of the original parent div width and height the size will still have a decrepancy. I will continue to work on this and send this website to people for their birthdays! Updating it regularly.